+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - [fix: tsf 2024 related bugfixes](https://github.com/Tencent/spring-cloud-tencent/pull/1664)
 - [fix: remove conflict dependencies management](https://github.com/Tencent/spring-cloud-tencent/pull/1669)
 - [refactor: optimize log](https://github.com/Tencent/spring-cloud-tencent/pull/1671)
+- [feat: support shortest response time lb and least connection lb](https://github.com/Tencent/spring-cloud-tencent/pull/1637)